Select all variables or the last variable

These functions are selection helpers.

everything() selects all variable. It is also useful in combination with other tidyselect operators.

last_col() selects the last variable.

everything(vars = peek_vars(fn = "everything"))

last_col(offset = 0L, vars = peek_vars(fn = "last_col"))

Arguments

vars
A character vector of variable names. If not supplied, the variables are taken from the current selection context (as established by functions like select() or pivot_longer()).

offset
Set it to n to select the nth var from the end.

Usage

Selection helpers can be used in functions like dplyr::select() or tidyr::pivot_longer(). Let's first attach the tidyverse:

library(tidyverse)

# For better printing
iris &lt;- as_tibble(iris)
mtcars &lt;- as_tibble(mtcars)

Use everything() to select all variables:

iris %&gt;% select(everything())

## # A tibble: 150 x 5
##   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
##          &lt;dbl&gt;       &lt;dbl&gt;        &lt;dbl&gt;       &lt;dbl&gt; &lt;fct&gt;  
## 1          5.1         3.5          1.4         0.2 setosa 
## 2          4.9         3            1.4         0.2 setosa 
## 3          4.7         3.2          1.3         0.2 setosa 
## 4          4.6         3.1          1.5         0.2 setosa 
## # ... with 146 more rows

mtcars %&gt;% pivot_longer(everything())

## # A tibble: 352 x 2
##   name  value
##   &lt;chr&gt; &lt;dbl&gt;
## 1 mpg      21
## 2 cyl       6
## 3 disp    160
## 4 hp      110
## # ... with 348 more rows

Use last_col() to select the last variable:

iris %&gt;% select(last_col())

## # A tibble: 150 x 1
##   Species
##   &lt;fct&gt;  
## 1 setosa 
## 2 setosa 
## 3 setosa 
## 4 setosa 
## # ... with 146 more rows

mtcars %&gt;% pivot_longer(last_col())

## # A tibble: 32 x 12
##     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear name  value
##   &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;chr&gt; &lt;dbl&gt;
## 1  21       6   160   110  3.9   2.62  16.5     0     1     4 carb      4
## 2  21       6   160   110  3.9   2.88  17.0     0     1     4 carb      4
## 3  22.8     4   108    93  3.85  2.32  18.6     1     1     4 carb      1
## 4  21.4     6   258   110  3.08  3.22  19.4     1     0     3 carb      1
## # ... with 28 more rows

Pass an offset to select a number of variables from the end:

mtcars %&gt;% select(1:last_col(5))

## # A tibble: 32 x 6
##     mpg   cyl  disp    hp  drat    wt
##   &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt;
## 1  21       6   160   110  3.9   2.62
## 2  21       6   160   110  3.9   2.88
## 3  22.8     4   108    93  3.85  2.32
## 4  21.4     6   258   110  3.08  3.22
## # ... with 28 more rows

See also