Selection language

Overview of selection features:

Tidyverse selections implement a dialect of R where operators make it easy to select variables:

: for selecting a range of consecutive variables.

! for taking the complement of a set of variables.

& and | for selecting the intersection or the union of two sets of variables.

c() for combining selections.

In addition, you can use selection helpers. Some helpers select specific columns:

everything(): Matches all variables.

last_col(): Select last variable, possibly with an offset.

These helpers select variables by matching patterns in their names:

starts_with(): Starts with a prefix.

ends_with(): Ends with a suffix.

contains(): Contains a literal string.

matches(): Matches a regular expression.

num_range(): Matches a numerical range like x01, x02, x03.

These helpers select variables from a character vector:

all_of(): Matches variable names in a character vector. All names must be present, otherwise an out-of-bounds error is thrown.

any_of(): Same as all_of(), except that no error is thrown for names that don't exist.

This helper selects variables with a function:

where(): Applies a function to all variables and selects those for which the function returns TRUE.

Simple examples

Here we show the usage for the basic selection operators. See the specific help pages to learn about helpers like starts_with().

The selection language can be used in functions like dplyr::select() or tidyr::pivot_longer(). Let's first attach the tidyverse:

library(tidyverse)

# For better printing
iris &lt;- as_tibble(iris)

Select variables by name:

starwars %&gt;% select(height)
#&gt; # A tibble: 87 x 1
#&gt;   height
#&gt;    &lt;int&gt;
#&gt; 1    172
#&gt; 2    167
#&gt; 3     96
#&gt; 4    202
#&gt; # ... with 83 more rows

iris %&gt;% pivot_longer(Sepal.Length)
#&gt; # A tibble: 150 x 6
#&gt;   Sepal.Width Petal.Length Petal.Width Species name         value
#&gt;         &lt;dbl&gt;        &lt;dbl&gt;       &lt;dbl&gt; &lt;fct&gt;   &lt;chr&gt;        &lt;dbl&gt;
#&gt; 1         3.5          1.4         0.2 setosa  Sepal.Length   5.1
#&gt; 2         3            1.4         0.2 setosa  Sepal.Length   4.9
#&gt; 3         3.2          1.3         0.2 setosa  Sepal.Length   4.7
#&gt; 4         3.1          1.5         0.2 setosa  Sepal.Length   4.6
#&gt; # ... with 146 more rows

Select multiple variables by separating them with commas. Note how the order of columns is determined by the order of inputs:

starwars %&gt;% select(homeworld, height, mass)
#&gt; # A tibble: 87 x 3
#&gt;   homeworld height  mass
#&gt;   &lt;chr&gt;      &lt;int&gt; &lt;dbl&gt;
#&gt; 1 Tatooine     172    77
#&gt; 2 Tatooine     167    75
#&gt; 3 Naboo         96    32
#&gt; 4 Tatooine     202   136
#&gt; # ... with 83 more rows

Functions like tidyr::pivot_longer() don't take variables with dots. In this case use c() to select multiple variables:

iris %&gt;% pivot_longer(c(Sepal.Length, Petal.Length))
#&gt; # A tibble: 300 x 5
#&gt;   Sepal.Width Petal.Width Species name         value
#&gt;         &lt;dbl&gt;       &lt;dbl&gt; &lt;fct&gt;   &lt;chr&gt;        &lt;dbl&gt;
#&gt; 1         3.5         0.2 setosa  Sepal.Length   5.1
#&gt; 2         3.5         0.2 setosa  Petal.Length   1.4
#&gt; 3         3           0.2 setosa  Sepal.Length   4.9
#&gt; 4         3           0.2 setosa  Petal.Length   1.4
#&gt; # ... with 296 more rows

Operators:

The : operator selects a range of consecutive variables:

starwars %&gt;% select(name:mass)
#&gt; # A tibble: 87 x 3
#&gt;   name           height  mass
#&gt;   &lt;chr&gt;           &lt;int&gt; &lt;dbl&gt;
#&gt; 1 Luke Skywalker    172    77
#&gt; 2 C-3PO             167    75
#&gt; 3 R2-D2              96    32
#&gt; 4 Darth Vader       202   136
#&gt; # ... with 83 more rows

The ! operator negates a selection:

starwars %&gt;% select(!(name:mass))
#&gt; # A tibble: 87 x 11
#&gt;   hair_color skin_color eye_color birth_year sex   gender homeworld species
#&gt;   &lt;chr&gt;      &lt;chr&gt;      &lt;chr&gt;          &lt;dbl&gt; &lt;chr&gt; &lt;chr&gt;  &lt;chr&gt;     &lt;chr&gt;  
#&gt; 1 blond      fair       blue            19   male  mascu~ Tatooine  Human  
#&gt; 2 &lt;NA&gt;       gold       yellow         112   none  mascu~ Tatooine  Droid  
#&gt; 3 &lt;NA&gt;       white, bl~ red             33   none  mascu~ Naboo     Droid  
#&gt; 4 none       white      yellow          41.9 male  mascu~ Tatooine  Human  
#&gt; # ... with 83 more rows, and 3 more variables: films &lt;list&gt;, vehicles &lt;list&gt;,
#&gt; #   starships &lt;list&gt;

iris %&gt;% select(!c(Sepal.Length, Petal.Length))
#&gt; # A tibble: 150 x 3
#&gt;   Sepal.Width Petal.Width Species
#&gt;         &lt;dbl&gt;       &lt;dbl&gt; &lt;fct&gt;  
#&gt; 1         3.5         0.2 setosa 
#&gt; 2         3           0.2 setosa 
#&gt; 3         3.2         0.2 setosa 
#&gt; 4         3.1         0.2 setosa 
#&gt; # ... with 146 more rows

iris %&gt;% select(!ends_with("Width"))
#&gt; # A tibble: 150 x 3
#&gt;   Sepal.Length Petal.Length Species
#&gt;          &lt;dbl&gt;        &lt;dbl&gt; &lt;fct&gt;  
#&gt; 1          5.1          1.4 setosa 
#&gt; 2          4.9          1.4 setosa 
#&gt; 3          4.7          1.3 setosa 
#&gt; 4          4.6          1.5 setosa 
#&gt; # ... with 146 more rows

& and | take the intersection or the union of two selections:

iris %&gt;% select(starts_with("Petal") &amp; ends_with("Width"))
#&gt; # A tibble: 150 x 1
#&gt;   Petal.Width
#&gt;         &lt;dbl&gt;
#&gt; 1         0.2
#&gt; 2         0.2
#&gt; 3         0.2
#&gt; 4         0.2
#&gt; # ... with 146 more rows

iris %&gt;% select(starts_with("Petal") | ends_with("Width"))
#&gt; # A tibble: 150 x 3
#&gt;   Petal.Length Petal.Width Sepal.Width
#&gt;          &lt;dbl&gt;       &lt;dbl&gt;       &lt;dbl&gt;
#&gt; 1          1.4         0.2         3.5
#&gt; 2          1.4         0.2         3  
#&gt; 3          1.3         0.2         3.2
#&gt; 4          1.5         0.2         3.1
#&gt; # ... with 146 more rows

To take the difference between two selections, combine the & and ! operators:

iris %&gt;% select(starts_with("Petal") &amp; !ends_with("Width"))
#&gt; # A tibble: 150 x 1
#&gt;   Petal.Length
#&gt;          &lt;dbl&gt;
#&gt; 1          1.4
#&gt; 2          1.4
#&gt; 3          1.3
#&gt; 4          1.5
#&gt; # ... with 146 more rows

Details

The order of selected columns is determined by the inputs.

all_of(c("foo", "bar")) selects "foo" first.

c(starts_with("c"), starts_with("d")) selects all columns starting with "c" first, then all columns starting with "d".