Superseded selection helpers

one_of() is superseded in favour of the more precise any_of() and all_of() selectors.

one_of(..., .vars = peek_vars(fn = "one_of"))

Arguments

...
One or more character vectors.

.vars
A character vector of variable names. When called from inside selecting functions like dplyr::select() these are automatically set to the names of the table.