Peek at variables in the selection context

peek_vars() returns the vector of names of the variables currently available for selection.

peek_data() returns the whole input vector (only available with eval_select()).

Read the Get started for examples of how to create selection helpers with peek_vars().

The variable names in a selection context are registered automatically by eval_select() and eval_rename() for the duration of the evaluation. peek_vars() is the glue that connects selection helpers to the current selection context.

peek_vars(..., fn = NULL)

peek_data(..., fn = NULL)

Arguments

...
These dots are for future extensions and must be empty.

fn
The name of the function to use in error messages when the helper is used in the wrong context. If not supplied, a generic error message is used instead.