tidyselect-package {tidyselect}	R Documentation

tidyselect: Select from a Set of Strings

Description

A backend for the selecting functions of the 'tidyverse'. It makes it easy to implement select-like functions in your own packages in a way that is consistent with other 'tidyverse' interfaces for selection.

Author(s)

Maintainer: Lionel Henry lionel@rstudio.com

Authors:

Hadley Wickham hadley@rstudio.com

Other contributors:

RStudio [copyright holder, funder]

See Also

Useful links:

https://tidyselect.r-lib.org

https://github.com/r-lib/tidyselect

Report bugs at https://github.com/r-lib/tidyselect/issues

[Package tidyselect version 1.2.0 Index]