The syntax and semantics of tidyselect

This is a technical description of the tidyselect interface.

Dots, c(), and unary -

tidyselect functions can take dots like dplyr::select(), or a named argument like tidyr::pivot_longer(). In the latter case, the dots syntax is accessible via c(). In fact ... syntax is implemented through c(...) and is thus completely equivalent.

# There is no semantical differences between these two expressions
dplyr::select(mtcars, mpg, disp:am)
dplyr::select(mtcars, c(mpg, disp:am))

Dots and c() are syntax for:

Set union

Set difference or complement, in combination with unary -

Renaming variables

Non-negative inputs are recursively joined with union(). The precedence is left-associative, just like with boolean operators. These expressions are all syntax for set union:

iris %>% select(starts_with("Sepal"), ends_with("Width"), Species)
iris %>% select(starts_with("Sepal") | ends_with("Width") | Species)
iris %>% select(union(union(starts_with("Sepal"), ends_with("Width")), 5L))

Unary - is normally syntax for set difference:

iris %>% select(starts_with("Sepal"), -ends_with("Width"), -Sepal.Length)
iris %>% select(setdiff(setdiff(starts_with("Sepal"), ends_with("Width")), 1L))

If the first ... or c() input is negative, an implicit everything() is appended. In this case, unary - is syntax for set complement.

iris %>% select(-starts_with("Sepal"))
iris %>% select(everything(), -starts_with("Sepal"))
iris %>% select(!starts_with("Sepal"))

If unary - is used outside ... or c(), it also stands for set complement.

Renaming variables

When named inputs are provided in ... or c(), the selection is renamed. If the inputs are already named, the outer and inner names are pasted together with a ... separator. Otherwise the outer names is propagated to the selected elements according to the following rules:

With data frames, a numeric suffix is appended because columns must be uniquely named.

With normal vectors, the name is simply assigned to all selected inputs.

Unnamed elements match any names:

a | c(foo = a) is equivalent to c(foo = a).

a & c(foo = a) is equivalent to c(foo = a).

c(foo = a) & c(bar = a) is equivalent to c().

c(foo = a) | c(bar = a) is equivalent to c(foo = a, bar = a).

Because unnamed elements match any named ones, it is possible to select multiple elements and rename one of them:

select(iris, starts_with("Sepal"), foo = Sepal.Width)