Select or rename variables

These functions power dplyr::select() and dplyr::rename(). They enable dplyr selecting or renaming semantics in your own functions.

vars_rename(.vars, ..., .strict = TRUE)

vars_select(.vars, ..., .include = character(), .exclude = character(),
  .strict = TRUE)

Arguments

.vars
A character vector of existing column names.

..., args
Selection inputs. See the help for selection helpers.

If you supply named inputs, the selected variables are renamed.

These arguments are automatically quoted and evaluated in a context where elements of vars are objects representing their positions within vars. They support unquoting and splicing. See vignette("programming") for an introduction to these concepts.

Note that except for :, - and c(), all complex expressions are evaluated outside that context. This is to prevent accidental matching to vars elements when you refer to variables from the calling context.

.strict
If TRUE, will throw an error if you attempt to select or rename a variable that doesn't exist.

.include, .exclude
Character vector of column names to always include/exclude.

Value

A named character vector. Values are existing column names, names are new names.

Customising error messages

For consistency with dplyr, error messages refer to "columns" by default. This assumes that the variables being selected come from a data frame. If this is not appropriate for your DSL, you can add an attribute type to the .vars vector to specify alternative names. This must be a character vector of length 2 whose first component is the singular form and the second is the plural. For example, c("variable", "variables").

Conditions

vars_select() signals the following warning.

tidyselect_warning_duplicate_renaming: Supplying named inputs in ... causes the variables to be renamed. For technical reasons there can only be one target name. If the same variable is renamed to different names, tidyselect issues this warning.

vars_select() signals the following conditions.

tidyselect_error_rename_to_same: Renaming multiple variables to the same name is an error.

tidyselect_error_rename_to_existing: Renaming a variable to an existing name is an error.

See also

Examples

# Keep variables
vars_select(names(iris), everything())

#>   Sepal.Length    Sepal.Width   Petal.Length    Petal.Width        Species 
#> "Sepal.Length"  "Sepal.Width" "Petal.Length"  "Petal.Width"      "Species" 

vars_select(names(iris), starts_with("Petal"))

#>   Petal.Length    Petal.Width 
#> "Petal.Length"  "Petal.Width" 

vars_select(names(iris), ends_with("Width"))

#>   Sepal.Width   Petal.Width 
#> "Sepal.Width" "Petal.Width" 

vars_select(names(iris), contains("etal"))

#>   Petal.Length    Petal.Width 
#> "Petal.Length"  "Petal.Width" 

vars_select(names(iris), matches(".t."))

#>   Sepal.Length    Sepal.Width   Petal.Length    Petal.Width 
#> "Sepal.Length"  "Sepal.Width" "Petal.Length"  "Petal.Width" 

vars_select(names(iris), Petal.Length, Petal.Width)

#>   Petal.Length    Petal.Width 
#> "Petal.Length"  "Petal.Width" 

vars_select(names(iris), one_of("Petal.Length", "Petal.Width"))

#>   Petal.Length    Petal.Width 
#> "Petal.Length"  "Petal.Width" 

df <- as.data.frame(matrix(runif(100), nrow = 10))
df <- df[c(3, 4, 7, 1, 9, 8, 5, 2, 6, 10)]
vars_select(names(df), num_range("V", 4:6))

#>   V4   V5   V6 
#> "V4" "V5" "V6" 

# Drop variables
vars_select(names(iris), -starts_with("Petal"))

#>   Sepal.Length    Sepal.Width        Species 
#> "Sepal.Length"  "Sepal.Width"      "Species" 

vars_select(names(iris), -ends_with("Width"))

#>   Sepal.Length   Petal.Length        Species 
#> "Sepal.Length" "Petal.Length"      "Species" 

vars_select(names(iris), -contains("etal"))

#>   Sepal.Length    Sepal.Width        Species 
#> "Sepal.Length"  "Sepal.Width"      "Species" 

vars_select(names(iris), -matches(".t."))

#>   Species 
#> "Species" 

vars_select(names(iris), -Petal.Length, -Petal.Width)

#>   Sepal.Length    Sepal.Width        Species 
#> "Sepal.Length"  "Sepal.Width"      "Species" 

# Rename variables
vars_select(names(iris), petal_length = Petal.Length)

#>   petal_length 
#> "Petal.Length" 

vars_select(names(iris), petal = starts_with("Petal"))

#>         petal1         petal2 
#> "Petal.Length"  "Petal.Width" 

# Rename variables preserving all existing
vars_rename(names(iris), petal_length = Petal.Length)

#>   Sepal.Length    Sepal.Width   petal_length    Petal.Width        Species 
#> "Sepal.Length"  "Sepal.Width" "Petal.Length"  "Petal.Width"      "Species" 

# You can unquote symbols or quosures
vars_select(names(iris), !! quote(Petal.Length))

#>   Petal.Length 
#> "Petal.Length" 

# And unquote-splice lists of symbols or quosures
vars_select(names(iris), !!! list(quo(Petal.Length), quote(Petal.Width)))

#>   Petal.Length    Petal.Width 
#> "Petal.Length"  "Petal.Width" 

# If you want to avoid ambiguity about where to find objects you
# have two solutions provided by the tidy eval framework. If you
# want to refer to local objects, you can explicitly unquote
# them. They must contain variable positions (integers) or variable
# names (strings):
Species <- 2
vars_select(names(iris), Species)     # Picks up `Species` from the data frame

#>   Species 
#> "Species" 

vars_select(names(iris), !! Species)  # Picks up the local object referring to column 2

#>   Sepal.Width 
#> "Sepal.Width" 

# If you want to make sure that a variable is picked up from the
# data, you can use the `.data` pronoun:
vars_select(names(iris), .data$Species)

#>   Species 
#> "Species" 

# If you're writing a wrapper around vars_select(), pass the dots
# via splicing to avoid matching dotted arguments to vars_select()
# named arguments (`vars`, `include` and `exclude`):
wrapper <- function(...) {
  vars_select(names(mtcars), !!! quos(...))
}

# This won't partial-match on `vars`:
wrapper(var = cyl)

#>   var 
#> "cyl" 

# This won't match on `include`:
wrapper(include = cyl)

#> include 
#>   "cyl" 

# If your wrapper takes named arguments, you need to capture then
# unquote to pass them to vars_select(). See the vignette on
# programming with dplyr for more on this:
wrapper <- function(var1, var2) {
  vars_select(names(mtcars), !! enquo(var1), !! enquo(var2))
}
wrapper(starts_with("d"), starts_with("c"))

#>   disp   drat    cyl   carb 
#> "disp" "drat"  "cyl" "carb"