Select all variables or the last variable

These functions are selection helpers.

everything() selects all variable. It is also useful in combination with other tidyselect operators.

last_col() selects the last variable.

everything(vars = NULL)

last_col(offset = 0L, vars = NULL)

Arguments

vars
A character vector of variable names. If not supplied, the variables are taken from the current selection context (as established by functions like select() or pivot_longer()).

offset
Set it to n to select the nth var from the end.

Examples

Selection helpers can be used in functions like dplyr::select() or tidyr::pivot_longer(). Let's first attach the tidyverse:

library(tidyverse)

# For better printing
iris &lt;- as_tibble(iris)
mtcars &lt;- as_tibble(mtcars)

Use everything() to select all variables:

iris %&gt;% select(everything())
#&gt; # A tibble: 150 x 5
#&gt;   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#&gt;          &lt;dbl&gt;       &lt;dbl&gt;        &lt;dbl&gt;       &lt;dbl&gt; &lt;fct&gt;  
#&gt; 1          5.1         3.5          1.4         0.2 setosa 
#&gt; 2          4.9         3            1.4         0.2 setosa 
#&gt; 3          4.7         3.2          1.3         0.2 setosa 
#&gt; 4          4.6         3.1          1.5         0.2 setosa 
#&gt; # ... with 146 more rows

mtcars %&gt;% pivot_longer(everything())
#&gt; # A tibble: 352 x 2
#&gt;   name  value
#&gt;   &lt;chr&gt; &lt;dbl&gt;
#&gt; 1 mpg      21
#&gt; 2 cyl       6
#&gt; 3 disp    160
#&gt; 4 hp      110
#&gt; # ... with 348 more rows

Use last_col() to select the last variable:

iris %&gt;% select(last_col())
#&gt; # A tibble: 150 x 1
#&gt;   Species
#&gt;   &lt;fct&gt;  
#&gt; 1 setosa 
#&gt; 2 setosa 
#&gt; 3 setosa 
#&gt; 4 setosa 
#&gt; # ... with 146 more rows

mtcars %&gt;% pivot_longer(last_col())
#&gt; # A tibble: 32 x 12
#&gt;     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear name  value
#&gt;   &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;chr&gt; &lt;dbl&gt;
#&gt; 1  21       6   160   110  3.9   2.62  16.5     0     1     4 carb      4
#&gt; 2  21       6   160   110  3.9   2.88  17.0     0     1     4 carb      4
#&gt; 3  22.8     4   108    93  3.85  2.32  18.6     1     1     4 carb      1
#&gt; 4  21.4     6   258   110  3.08  3.22  19.4     1     0     3 carb      1
#&gt; # ... with 28 more rows

Supply an offset n to select a variable located n positions from the end:

mtcars %&gt;% select(1:last_col(5))
#&gt; # A tibble: 32 x 6
#&gt;     mpg   cyl  disp    hp  drat    wt
#&gt;   &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt;
#&gt; 1  21       6   160   110  3.9   2.62
#&gt; 2  21       6   160   110  3.9   2.88
#&gt; 3  22.8     4   108    93  3.85  2.32
#&gt; 4  21.4     6   258   110  3.08  3.22
#&gt; # ... with 28 more rows

See also