FAQ - Note: Using an external vector in selections is ambiguous

Ambiguity between columns and external variables

With selecting functions like dplyr::select() or tidyr::pivot_longer(), you can refer to variables by name:

mtcars %&gt;% select(cyl, am, vs)
#&gt; # A tibble: 32 x 3
#&gt;     cyl    am    vs
#&gt;   &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt;
#&gt; 1     6     1     0
#&gt; 2     6     1     0
#&gt; 3     4     1     1
#&gt; 4     6     0     1
#&gt; # ... with 28 more rows

mtcars %&gt;% select(mpg:disp)
#&gt; # A tibble: 32 x 3
#&gt;     mpg   cyl  disp
#&gt;   &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt;
#&gt; 1  21       6   160
#&gt; 2  21       6   160
#&gt; 3  22.8     4   108
#&gt; 4  21.4     6   258
#&gt; # ... with 28 more rows

For historical reasons, it is also possible to refer an external vector of variable names. You get the correct result, but with a note informing you that selecting with an external variable is ambiguous because it is not clear whether you want a data frame column or an external object.

vars &lt;- c("cyl", "am", "vs")
result &lt;- mtcars %&gt;% select(vars)
#&gt; Note: Using an external vector in selections is ambiguous.
#&gt; i Use `all_of(vars)` instead of `vars` to silence this message.
#&gt; i See &lt;https://tidyselect.r-lib.org/reference/faq-external-vector.html&gt;.
#&gt; This message is displayed once per session.

This note will become a warning in the future, and then an error. We have decided to deprecate this particular approach to using external vectors because they introduce ambiguity. Imagine that the data frame contains a column with the same name as your external variable.

some_df &lt;- mtcars[1:4, ]
some_df$vars &lt;- 1:nrow(some_df)

These are very different objects but it isn’t a problem if the context forces you to be specific about where to find vars:

vars
#&gt; [1] "cyl" "am"  "vs"

some_df$vars
#&gt; [1] 1 2 3 4

In a selection context however, the column wins:

some_df %&gt;% select(vars)
#&gt; # A tibble: 4 x 1
#&gt;    vars
#&gt;   &lt;int&gt;
#&gt; 1     1
#&gt; 2     2
#&gt; 3     3
#&gt; 4     4

Fixing the ambiguity

To make your selection code more robust and silence the message, use all_of() to force the external vector:

some_df %&gt;% select(all_of(vars))
#&gt; # A tibble: 4 x 3
#&gt;     cyl    am    vs
#&gt;   &lt;dbl&gt; &lt;dbl&gt; &lt;dbl&gt;
#&gt; 1     6     1     0
#&gt; 2     6     1     0
#&gt; 3     4     1     1
#&gt; 4     6     0     1

For more information or if you have comments about this, please see the Github issue tracking the deprecation process.