one_of {tidyselect}	R Documentation

Superseded selection helpers

Description

one_of() is superseded in favour of the more precise any_of() and all_of() selectors.

Usage

one_of(..., .vars = NULL)

Arguments

...
One or more character vectors.

.vars
A character vector of variable names. When called from inside selecting functions like dplyr::select() these are automatically set to the names of the table.

[Package tidyselect version 1.2.0 Index]