select_helpers {tidyselect}	R Documentation

Select helpers

Description

These functions allow you to select variables based on their names.

starts_with(): Starts with a prefix.

ends_with(): Ends with a suffix.

contains(): Contains a literal string.

matches(): Matches a regular expression.

num_range(): Matches a numerical range like x01, x02, x03.

one_of(): Matches variable names in a character vector.

everything(): Matches all variables.

last_col(): Select last variable, possibly with an offset.

Usage

starts_with(match, ignore.case = TRUE, vars = peek_vars())

ends_with(match, ignore.case = TRUE, vars = peek_vars())

contains(match, ignore.case = TRUE, vars = peek_vars())

matches(match, ignore.case = TRUE, vars = peek_vars())

num_range(prefix, range, width = NULL, vars = peek_vars())

one_of(..., .vars = peek_vars())

everything(vars = peek_vars())

last_col(offset = 0L, vars = peek_vars())

Arguments

match
A string.

ignore.case
If TRUE, the default, ignores case when matching names.

vars, .vars
A character vector of variable names. When called from inside selecting functions like dplyr::select() these are automatically set to the names of the table.

prefix
A prefix that starts the numeric range.

range
A sequence of integers, like 1:5.

width
Optionally, the "width" of the numeric range. For example, a range of 2 gives "01", a range of three "001", etc.

...
One or more character vectors.

offset
Set it to n to select the nth var from the end.

Value

An integer vector giving the position of the matched variables.

Examples

nms <- names(iris)
vars_select(nms, starts_with("Petal"))
vars_select(nms, ends_with("Width"))
vars_select(nms, contains("etal"))
vars_select(nms, matches(".t."))
vars_select(nms, Petal.Length, Petal.Width)
vars_select(nms, everything())
vars_select(nms, last_col())
vars_select(nms, last_col(offset = 2))

vars <- c("Petal.Length", "Petal.Width")
vars_select(nms, one_of(vars))

[Package tidyselect version 0.2.5 Index]