Select or rename variables (questioning)

These functions are now in the questioning stage. They might be retired or deprecated in the future.

Please use eval_select() and eval_rename() instead. See vignette("tidyselect") to get started.

vars_select(
  .vars,
  ...,
  .include = character(),
  .exclude = character(),
  .strict = TRUE
)

vars_rename(.vars, ..., .strict = TRUE)

Arguments

.vars
A character vector of existing column names.

..., args
Selection inputs. See the help for selection helpers.

.include, .exclude
Character vector of column names to always include/exclude.

.strict
If TRUE, will throw an error if you attempt to select or rename a variable that doesn't exist.

Value

A named character vector. Values are existing column names, names are new names.

See also