Select variables with a function

This selection helper selects the variables for which a function returns TRUE.

where(fn)

Arguments

fn
A function that returns TRUE or FALSE (technically, a predicate function). Can also be a purrr-like formula.

Examples

Selection helpers can be used in functions like dplyr::select() or tidyr::pivot_longer(). Let's first attach the tidyverse:

library(tidyverse)

# For better printing
iris &lt;- as_tibble(iris)

where() takes a function and returns all variables for which the function returns TRUE:

is.factor(iris[[4]])
#&gt; [1] FALSE

is.factor(iris[[5]])
#&gt; [1] TRUE

iris %&gt;% select(where(is.factor))
#&gt; # A tibble: 150 x 1
#&gt;   Species
#&gt;   &lt;fct&gt;  
#&gt; 1 setosa 
#&gt; 2 setosa 
#&gt; 3 setosa 
#&gt; 4 setosa 
#&gt; # ... with 146 more rows

is.numeric(iris[[4]])
#&gt; [1] TRUE

is.numeric(iris[[5]])
#&gt; [1] FALSE

iris %&gt;% select(where(is.numeric))
#&gt; # A tibble: 150 x 4
#&gt;   Sepal.Length Sepal.Width Petal.Length Petal.Width
#&gt;          &lt;dbl&gt;       &lt;dbl&gt;        &lt;dbl&gt;       &lt;dbl&gt;
#&gt; 1          5.1         3.5          1.4         0.2
#&gt; 2          4.9         3            1.4         0.2
#&gt; 3          4.7         3.2          1.3         0.2
#&gt; 4          4.6         3.1          1.5         0.2
#&gt; # ... with 146 more rows

The formula shorthand

You can use purrr-like formulas as a shortcut for creating a function on the spot. These expressions are equivalent:

iris %&gt;% select(where(is.numeric))
#&gt; # A tibble: 150 x 4
#&gt;   Sepal.Length Sepal.Width Petal.Length Petal.Width
#&gt;          &lt;dbl&gt;       &lt;dbl&gt;        &lt;dbl&gt;       &lt;dbl&gt;
#&gt; 1          5.1         3.5          1.4         0.2
#&gt; 2          4.9         3            1.4         0.2
#&gt; 3          4.7         3.2          1.3         0.2
#&gt; 4          4.6         3.1          1.5         0.2
#&gt; # ... with 146 more rows

iris %&gt;% select(where(function(x) is.numeric(x)))
#&gt; # A tibble: 150 x 4
#&gt;   Sepal.Length Sepal.Width Petal.Length Petal.Width
#&gt;          &lt;dbl&gt;       &lt;dbl&gt;        &lt;dbl&gt;       &lt;dbl&gt;
#&gt; 1          5.1         3.5          1.4         0.2
#&gt; 2          4.9         3            1.4         0.2
#&gt; 3          4.7         3.2          1.3         0.2
#&gt; 4          4.6         3.1          1.5         0.2
#&gt; # ... with 146 more rows

iris %&gt;% select(where(~ is.numeric(.x)))
#&gt; # A tibble: 150 x 4
#&gt;   Sepal.Length Sepal.Width Petal.Length Petal.Width
#&gt;          &lt;dbl&gt;       &lt;dbl&gt;        &lt;dbl&gt;       &lt;dbl&gt;
#&gt; 1          5.1         3.5          1.4         0.2
#&gt; 2          4.9         3            1.4         0.2
#&gt; 3          4.7         3.2          1.3         0.2
#&gt; 4          4.6         3.1          1.5         0.2
#&gt; # ... with 146 more rows

The shorthand is useful for adding logic inline. Here we select all numeric variables whose mean is greater than 3.5:

iris %&gt;% select(where(~ is.numeric(.x) &amp;&amp; mean(.x) &gt; 3.5))
#&gt; # A tibble: 150 x 2
#&gt;   Sepal.Length Petal.Length
#&gt;          &lt;dbl&gt;        &lt;dbl&gt;
#&gt; 1          5.1          1.4
#&gt; 2          4.9          1.4
#&gt; 3          4.7          1.3
#&gt; 4          4.6          1.5
#&gt; # ... with 146 more rows